+++ SocialChoring/src/com/socialchoring/bean/PlayerGames.java
@@ -4,31 +4,31 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement
 public class PlayerGames {
-	private int id;
-	private int game_id;
-	private int player_id;
+	private long id;
+	private long game_id;
+	private long player_id;
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 
-	public int getGame_id() {
+	public long getGame_id() {
 		return this.game_id;
 	}
 
-	public void setGame_id(int game_id) {
+	public void setGame_id(long game_id) {
 		this.game_id = game_id;
 	}
 
-	public int getPlayer_id() {
+	public long getPlayer_id() {
 		return this.player_id;
 	}
 
-	public void setPlayer_id(int player_id) {
+	public void setPlayer_id(long player_id) {
 		this.player_id = player_id;
 	}
